Moved all macros into a single macro.rs file at the root of the library.
